requirements.txt
Django
djangorestframework
django-npm
cryptography
